Fix #6 - Add support for AWS.Credentials objects
Should be backwards compatible: if for some reason you have an object that has both the old style and new (standard)
style creds properties, the old-style will be used. Only possible break in compatibility is if someone for some
reason passed in options that have the new style credential properties but not the old ones, and was relying
on the previous behavior to load the creds from env vars: now, the new-style credentials will be favored if
present over the env vars.